+++ action-sheet/android/src/main/java/com/capacitorjs/plugins/actionsheet/ActionSheet.java
@@ -26,7 +26,9 @@ public class ActionSheet extends BottomSheetDialogFragment {
     @Override
     public void onCancel(DialogInterface dialog) {
         super.onCancel(dialog);
-        this.cancelListener.onCancel();
+        if (this.cancelListener != null) {
+            this.cancelListener.onCancel();
+        }
     }
 
     private String title;
@@ -101,15 +103,13 @@ public class ActionSheet extends BottomSheetDialogFragment {
             tv.setTextColor(Color.parseColor("#000000"));
             tv.setPadding(layoutPaddingPx12, layoutPaddingPx12, layoutPaddingPx12, layoutPaddingPx12);
             tv.setText(options[i].getTitle());
-            tv.setOnClickListener(
-                view -> {
-                    Logger.debug("CliCKED: " + optionIndex);
+            tv.setOnClickListener((view) -> {
+                Logger.debug("CliCKED: " + optionIndex);
 
-                    if (listener != null) {
-                        listener.onSelect(optionIndex);
-                    }
+                if (listener != null) {
+                    listener.onSelect(optionIndex);
                 }
-            );
+            });
             layout.addView(tv);
         }