+++ action-sheet/ios/Sources/ActionSheetPlugin/ActionSheetPlugin.swift
@@ -6,13 +6,14 @@ import Capacitor
  * here: https://capacitorjs.com/docs/plugins/ios
  */
 @objc(ActionSheetPlugin)
-public class ActionSheetPlugin: CAPPlugin, CAPBridgedPlugin {
+public class ActionSheetPlugin: CAPPlugin, CAPBridgedPlugin, UIAdaptivePresentationControllerDelegate {
     public let identifier = "ActionSheetPlugin"
     public let jsName = "ActionSheet"
     public let pluginMethods: [CAPPluginMethod] = [
         CAPPluginMethod(name: "showActions", returnType: CAPPluginReturnPromise)
     ]
     private let implementation = ActionSheet()
+    private var currentCall: CAPPluginCall?
 
     @objc func showActions(_ call: CAPPluginCall) {
         let title = call.options["title"] as? String
@@ -20,6 +21,7 @@ public class ActionSheetPlugin: CAPPlugin, CAPBridgedPlugin {
 
         let options = call.getArray("options", JSObject.self) ?? []
         var alertActions = [UIAlertAction]()
+        var hasCancellableButton = false
         for (index, option) in options.enumerated() {
             let style = option["style"] as? String ?? "DEFAULT"
             let title = option["title"] as? String ?? ""
@@ -27,12 +29,15 @@ public class ActionSheetPlugin: CAPPlugin, CAPBridgedPlugin {
             if style == "DESTRUCTIVE" {
                 buttonStyle = .destructive
             } else if style == "CANCEL" {
+                hasCancellableButton = true
                 buttonStyle = .cancel
             }
-            let action = UIAlertAction(title: title, style: buttonStyle, handler: { (_) -> Void in
+            let action = UIAlertAction(title: title, style: buttonStyle, handler: { [weak self] (_) in
                 call.resolve([
-                    "index": index
+                    "index": index,
+                    "canceled": buttonStyle == .cancel
                 ])
+                self?.currentCall = nil
             })
             alertActions.append(action)
         }
@@ -40,9 +45,61 @@ public class ActionSheetPlugin: CAPPlugin, CAPBridgedPlugin {
         DispatchQueue.main.async { [weak self] in
             if let alertController = self?.implementation.buildActionSheet(title: title, message: message, actions: alertActions) {
                 self?.setCenteredPopover(alertController)
-                self?.bridge?.viewController?.present(alertController, animated: true, completion: nil)
+                self?.bridge?.viewController?.present(alertController, animated: true) {
+                    if !hasCancellableButton {
+                        self?.setupCancelationListeners(alertController, call)
+                    }
+                }
             }
         }
     }
 
+    private func setupCancelationListeners(_ alertController: UIAlertController, _ call: CAPPluginCall) {
+        let cancelable = call.getBool("cancelable", false)
+        if #available(iOS 26, *) {
+            self.currentCall = call
+            alertController.presentationController?.delegate = self
+        } else if cancelable {
+            // For iOS versions below 26, setting the presentation controller delegate would result in a crash
+            //  "Terminating app due to uncaught exception 'NSInternalInconsistencyException', reason: 'The presentation controller of an alert controller presenting as an alert must not have its delegate modified"
+            //  Hence, the alternative by adding a gesture recognizer (which only works for iOS versions below 26)
+            let gestureRecognizer = TapGestureRecognizerWithClosure {
+                alertController.dismiss(animated: true, completion: nil)
+                call.actionSheetCanceled()
+            }
+            let backgroundView = alertController.view.superview?.subviews[0]
+            backgroundView?.addGestureRecognizer(gestureRecognizer)
+        }
+    }
+
+    // MARK: - UIAdaptivePresentationControllerDelegate
+
+    public func presentationControllerDidDismiss(_ presentationController: UIPresentationController) {
+        self.currentCall?.actionSheetCanceled()
+        self.currentCall = nil
+    }
+}
+
+// MARK: - TapGestureRecognizerWithClosure
+private final class TapGestureRecognizerWithClosure: UITapGestureRecognizer {
+    private let onTap: () -> Void
+
+    init(onTap: @escaping () -> Void) {
+        self.onTap = onTap
+        super.init(target: nil, action: nil)
+        self.addTarget(self, action: #selector(action))
+    }
+
+    @objc private func action() {
+        onTap()
+    }
+}
+
+private extension CAPPluginCall {
+    func actionSheetCanceled() {
+        resolve([
+            "index": -1,
+            "canceled": true
+        ])
+    }
 }