+++ local-notifications/src/index.ts
@@ -2,12 +2,9 @@ import { registerPlugin } from '@capacitor/core';
 
 import type { LocalNotificationsPlugin } from './definitions';
 
-const LocalNotifications = registerPlugin<LocalNotificationsPlugin>(
-  'LocalNotifications',
-  {
-    web: () => import('./web').then(m => new m.LocalNotificationsWeb()),
-  },
-);
+const LocalNotifications = registerPlugin<LocalNotificationsPlugin>('LocalNotifications', {
+  web: () => import('./web').then((m) => new m.LocalNotificationsWeb()),
+});
 
 export * from './definitions';
 export { LocalNotifications };