+++ status-bar/src/definitions.ts
@@ -1,5 +1,7 @@
 /// <reference types="@capacitor/cli" />
 
+import type { PluginListenerHandle } from '@capacitor/core';
+
 declare module '@capacitor/cli' {
   export interface PluginsConfig {
     /**
@@ -8,10 +10,7 @@ declare module '@capacitor/cli' {
     StatusBar?: {
       /**
        * Whether the statusbar is overlaid or not.
-       * For applications targeting Android 15, this property has no effect unless
-       * the property windowOptOutEdgeToEdgeEnforcement is added to the application layout file.
-       * Otherwise, the application assumes always overlays as true.
-       * More details in https://developer.android.com/reference/android/R.attr#windowOptOutEdgeToEdgeEnforcement
+       * Not available on Android 15+.
        *
        * @since 1.0.0
        * @default true
@@ -31,6 +30,7 @@ declare module '@capacitor/cli' {
       /**
        * Color of the background of the statusbar in hex format, #RRGGBB.
        * Doesn't work if `overlaysWebView` is true.
+       * Not available on Android 15+.
        *
        * @since 1.0.0
        * @default #000000
@@ -143,14 +143,21 @@ export interface StatusBarInfo {
    *
    * @since 1.0.0
    */
-  color?: string;
+  color: string;
 
   /**
-   * Whether the statusbar is overlaid or not.
+   * Whether the status bar is overlaid or not.
    *
    * @since 1.0.0
    */
-  overlays?: boolean;
+  overlays: boolean;
+
+  /**
+   * The height of the status bar.
+   *
+   * @since 7.0.0
+   */
+  height: number;
 }
 
 export interface SetOverlaysWebViewOptions {
@@ -162,6 +169,9 @@ export interface SetOverlaysWebViewOptions {
   overlay: boolean;
 }
 
+export type VisibilityChangeListener = (info: StatusBarInfo) => void;
+export type OverlayChangeListener = (info: StatusBarInfo) => void;
+
 export interface StatusBarPlugin {
   /**
    * Set the current style of the status bar.
@@ -172,6 +182,8 @@ export interface StatusBarPlugin {
 
   /**
    * Set the background color of the status bar.
+   * Calling this function updates the foreground color of the status bar if the style is set to default, except on iOS versions lower than 17.
+   * Not available on Android 15+.
    *
    * @since 1.0.0
    */
@@ -205,10 +217,30 @@ export interface StatusBarPlugin {
   /**
    * Set whether or not the status bar should overlay the webview to allow usage
    * of the space underneath it.
+   * Not available on Android 15+.
    *
    * @since 1.0.0
    */
   setOverlaysWebView(options: SetOverlaysWebViewOptions): Promise<void>;
+
+  /**
+   * Listen for status bar visibility changes.
+   * Fired when hide or show methods get called.
+   *
+   * @since 7.0.0
+   */
+  addListener(
+    eventName: 'statusBarVisibilityChanged',
+    listenerFunc: VisibilityChangeListener,
+  ): Promise<PluginListenerHandle>;
+
+  /**
+   * Listen for status bar overlay changes.
+   * Fired when setOverlaysWebView gets called.
+   *
+   * @since 7.0.0
+   */
+  addListener(eventName: 'statusBarOverlayChanged', listenerFunc: OverlayChangeListener): Promise<PluginListenerHandle>;
 }
 
 /**